Add custom translation function
In order to get the translation from the DB, a new
translate function has been added, called `g`. This function
works like the `I18n.t`, but looking for the string in the
DB instead of on the yml files.

The t('...') function has been changed for g('...') in views and controllers.