Fix spec for uniqness of a budget group name
This spec[1] was failing because we were adding the error message to the `budget_group` itself, but it should be added to the Translation object for the `budget_group`, as explaned in this commit[2].

Note: this is not a production ready solution. We should try to figure out how to add it to the correct traslation object, not the `last` translation. And it also seems strange to have to add a doble error message: One for the `budget_group` itself, so that the `budget_group` is not saved, and another one for the translation object to display the error in the interface.

We'll figure it out :relieved:

[1] https://github.com/AyuntamientoMadrid/consul/pull/1842/files#diff-f61d83c6f77178d1380597cbae2f6490R194

[2] https://github.com/rockandror/consul/commit/dca95d608f898c3ef95c98af4a66fc65ec5bf723