Fix spec for updating translation of a budget group
This spec[1] was failing due to a `before_save` method in the model, which was changing the value of an attribute in the default locale.

So, eventhough the intend of the spec was correct: Deleting the default locale translation. As we were changing the default translations before saving it, the translation for the default locale  existed and thus was not deleted. Well actually, looking at the sql, it was deleted and then created again.

By using `name.strip!` instead of `self.name = self.name.strip` the spec passes again.

Note: This is not ready for production. There is no model spec in the budget_group_spec.rb to test this. I'm 100% sure that it will not have side effects without specs.

[1] Update a translation not having the current locale
spec/features/admin/budget_groups_spec.rb[1:1:1:5]