Update specs
We now have a dependency on this spec, added in a previous PR[1].

This is due the spec indirectly creating a Comment object[1] that requires the spending proposal to have an administrator.

We could separate the rake task to deal with the initial PR requirements and this PR's requirements. But we should put it all in a single rake task when everything is ready. So just fixing the spec for now.

[1] https://github.com/AyuntamientoMadrid/consul/pull/1853