Migrate ballot lines in all of a user's ballots
Some users have multiple ballots with different ballot lines.

This is due to reclasifications process that was done during the voting phase, were spending proposals were move from one geozone to another one. Creating a second ballot for the spending proposal's reclassified geozone.

With this commit we are migrating all ballot lines of these ballots, to a single ballot.

Note that this change is related to the scenario where a user has delegated and also voted. In that case, the migration should not add the ballot lines from the delegated representative[1].

[1] https://github.com/AyuntamientoMadrid/consul/pull/1776/files#diff-ed95aefdf8cdd908de495dfc558773beR172